# Optimize adding a single component type to an entity, add benchmarks