--- a/src/Arch.Benchmarks/ArchetypeChangeBenchmark.cs
+++ b/src/Arch.Benchmarks/ArchetypeChangeBenchmark.cs
@@ -8,33 +8,49 @@ namespace Arch.Benchmarks;
 // [HardwareCounters(HardwareCounter.CacheMisses)]
 public class ArchetypeChangeBenchmark
 {
-    private readonly ComponentType[] _group = { typeof(Transform), typeof(Rotation) };
+    private readonly ComponentType[] _groupOne = { typeof(Transform) };
+
+    private readonly ComponentType[] _groupFive =
+    {
+        typeof(Transform), typeof(Velocity), typeof(Rotation), typeof(Position2D), typeof(Position3D)
+    };
 
     private World _world = default!;
 
     [Params(10000, 100000, 1000000)] public int Amount = default!;
 
-    private Entity[] Entities = default!;
+    private Entity[] _entitiesOne = default!;
+    private Entity[] _entitiesFive = default!;
 
     [GlobalSetup]
     public void Setup()
     {
         _world = World.Create();
-        Entities = new Entity[Amount];
+        _entitiesOne = new Entity[Amount];
+        _entitiesFive = new Entity[Amount];
 
         for (var i = 0; i < Amount; i++)
         {
-            _world.Create(_group);
-            Entities[i] = new Entity(i, _world.Id);
+            _entitiesOne[i] = _world.Create(_groupOne);
+            _entitiesFive[i] = _world.Create(_groupFive);
         }
     }
 
     [Benchmark]
-    public void ChangeArchetypeTwoToThree()
+    public void AddArchetypeOne()
     {
-        foreach (var entity in Entities.AsSpan())
+        foreach (var entity in _entitiesOne.AsSpan())
         {
-            _world.Add<Velocity>(entity);
+            _world.Add<Health>(entity);
+        }
+    }
+
+    [Benchmark]
+    public void AddArchetypeFive()
+    {
+        foreach (var entity in _entitiesOne.AsSpan())
+        {
+            _world.Add<Health>(entity);
         }
     }
 }