--- a/src/Arch.Benchmarks/Utils/Structs.cs
+++ b/src/Arch.Benchmarks/Utils/Structs.cs
@@ -15,4 +15,19 @@ public struct Rotation
     public float X, Y, Z, W;
 }
 
+public struct Position2D
+{
+    public float X, Y;
+}
+
+public struct Position3D
+{
+    public float X, Y, Z;
+}
+
+public struct Health
+{
+    public int Amount;
+}
+
 // public struct Ai { }