Also makes Chunk.Has and Chunk.CopyComponents faster.
Heavily inspired by flecs. See https://github.com/SanderMertens/flecs/blob/master/docs/Relationships.md#id-ranges and https://ajmmertens.medium.com/building-an-ecs-2-archetypes-and-vectorization-fe21690805f9
When the component type id is below 256 an array is used, otherwise a dictionary is used, similar to flecs.
The array is only allocated after the first time that an archetype is modified, and only to the size that it needs to fit the component type id with each usage.
A preprocessor directive is used in case CollectionsMarshal does not exist (pre .NET 5)

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.2846)
AMD Ryzen 9 7900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.203
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2

---

Old:
|           Method |  Amount |       Mean |     Error |    StdDev |     Median | Allocated |
|----------------- |-------- |-----------:|----------:|----------:|-----------:|----------:|
|  AddArchetypeOne |   10000 |   1.409 ms | 0.0280 ms | 0.0526 ms |   1.387 ms |       2 B |
| AddArchetypeFive |   10000 |   3.165 ms | 0.0084 ms | 0.0070 ms |   3.165 ms |       3 B |
|  AddArchetypeOne |  100000 |  13.738 ms | 0.1541 ms | 0.1442 ms |  13.808 ms |      13 B |
| AddArchetypeFive |  100000 |  31.367 ms | 0.0547 ms | 0.0427 ms |  31.373 ms |      51 B |
|  AddArchetypeOne | 1000000 | 135.100 ms | 0.3742 ms | 0.3500 ms | 135.049 ms |     204 B |
| AddArchetypeFive | 1000000 | 313.442 ms | 0.8176 ms | 0.7648 ms | 313.690 ms |     816 B |

---

New:
|           Method |  Amount |         Mean |       Error |      StdDev | Allocated |
|----------------- |-------- |-------------:|------------:|------------:|----------:|
|  AddArchetypeOne |   10000 |     571.5 us |     6.90 us |     5.76 us |       1 B |
| AddArchetypeFive |   10000 |   1,416.0 us |     7.15 us |     5.58 us |       2 B |
|  AddArchetypeOne |  100000 |   5,651.4 us |    10.90 us |     9.66 us |       6 B |
| AddArchetypeFive |  100000 |  13,616.9 us |    48.06 us |    42.60 us |      13 B |
|  AddArchetypeOne | 1000000 |  56,841.2 us |   185.63 us |   155.01 us |      91 B |
| AddArchetypeFive | 1000000 | 139,875.7 us | 2,063.16 us | 1,929.88 us |     204 B |