@DrSmugleaf 2023-05-08
As a comparison, these are the same benchmarks without the World.Add or Archetype changes, using only the changes in Chunk.cs:
|           Method |  Amount |         Mean |       Error |      StdDev | Allocated |
|----------------- |-------- |-------------:|------------:|------------:|----------:|
|  AddArchetypeOne |   10000 |     862.6 us |    16.83 us |    16.53 us |       1 B |
| AddArchetypeFive |   10000 |   1,814.8 us |     4.53 us |     3.78 us |       2 B |
|  AddArchetypeOne |  100000 |   8,507.4 us |   131.22 us |   122.74 us |      13 B |
| AddArchetypeFive |  100000 |  17,359.0 us |   304.09 us |   284.45 us |      26 B |
|  AddArchetypeOne | 1000000 |  86,454.2 us |   875.31 us |   818.77 us |     117 B |
| AddArchetypeFive | 1000000 | 175,630.9 us | 3,484.25 us | 5,916.53 us |     272 B |