@genaray 2023-05-08
Looks great! :) 
Just out of curiosity, is it possible to create its own container/data class that switches between dictionary and array based on its element amount? 

E.g. i just realize that there other aspects which could make use of a similar mechanic ^^