@DrSmugleaf 2023-05-08
> Looks great! :) Just out of curiosity, is it possible to create its own container/data class that switches between dictionary and array based on its element amount?
> 
> E.g. i just realize that there other aspects which could make use of a similar mechanic ^^

Done in [this commit](https://github.com/genaray/Arch/pull/103/commits/d8be5cf3145b1e47daf8720bf664e8b18bc29414)

New benchmarks:
BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19045.2846)
AMD Ryzen 9 7900X, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.203
  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2

---

Old:
|           Method |  Amount |         Mean |       Error |      StdDev |       Median | Allocated |
|----------------- |-------- |-------------:|------------:|------------:|-------------:|----------:|
|  AddArchetypeOne |   10000 |     589.0 us |    10.90 us |    22.02 us |     579.9 us |       1 B |
| AddArchetypeFive |   10000 |   1,329.4 us |     3.61 us |     3.01 us |   1,328.4 us |       2 B |
|  AddArchetypeOne |  100000 |   5,649.0 us |    13.75 us |    11.48 us |   5,648.7 us |       6 B |
| AddArchetypeFive |  100000 |  13,768.5 us |    54.73 us |    48.52 us |  13,752.0 us |      13 B |
|  AddArchetypeOne | 1000000 |  55,089.3 us |   246.16 us |   205.56 us |  55,104.9 us |      82 B |
| AddArchetypeFive | 1000000 | 141,392.8 us | 3,213.12 us | 9,473.95 us | 138,269.6 us |     204 B |

---

New:
|           Method |  Amount |         Mean |     Error |    StdDev | Allocated |
|----------------- |-------- |-------------:|----------:|----------:|----------:|
|  AddArchetypeOne |   10000 |     575.0 us |   0.80 us |   0.71 us |       1 B |
| AddArchetypeFive |   10000 |   1,378.2 us |   3.25 us |   2.54 us |       2 B |
|  AddArchetypeOne |  100000 |   5,787.7 us |  38.90 us |  32.49 us |       6 B |
| AddArchetypeFive |  100000 |  13,590.3 us |  32.61 us |  28.91 us |      13 B |
|  AddArchetypeOne | 1000000 |  58,872.8 us | 208.38 us | 184.73 us |      91 B |
| AddArchetypeFive | 1000000 | 141,110.8 us | 432.42 us | 404.48 us |     204 B |